requirements.txt
# Core AutoGen packages
pyautogen
openai
python-dotenv
sentence-transformers
pinecone